Lecture-to-notes/requirements.txt
streamlit
openai-whisper
transformers
torch
fpdf
google-genai
python-dotenv
imageio[ffmpeg]
